Add ideas.md, move models, refactor and delete old ideas
