+++ projects/ngx-testing-extra-demo/src/app/components/page-2/page-2.component.spec.ts
@@ -2,30 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ExtraBed, ExtraFn } from 'ngx-testing-extra';
 import { Page2Component } from './page-2.component';
 
-const mockWindow = { back: () => {} };
-
-// describeComponent(Page2Component, ({ bed }) => {
-//   bed.provide([{ provide: WINDOW, useValue: mockWindow }]);
-//
-//   bed.compileBeforeEach(({ instance }) => {
-//
-//     it('should create', () => {
-//       expect(instance).toBeTruthy();
-//     });
-//
-//     it('should create', () => {
-//       expect(instance.checked).toBeFalse();
-//       instance.checked = true;
-//       expect(instance.checked).toBeTrue();
-//     });
-//
-//     it('should ', () => {
-//       expect(instance.checked).toBeFalse();
-//     });
-//   });
-// });
-
-describe('Page2Component', () => {
+xdescribe('Page2Component', () => {
   let extra: ExtraFn<Page2Component>;
 
   beforeEach(async () => {
@@ -37,7 +14,7 @@ describe('Page2Component', () => {
     expect(instance.checked).toBeFalse();
     instance.checked = true;
     expect(instance.checked).toBeTrue();
-  }, {}));
+  }));
 
   it('should ', extra(({ instance }) => {
     expect(instance.checked).toBeFalse();
