+++ projects/ngx-testing-extra/src/lib/core/expect/should-create.ts
@@ -1,0 +1,3 @@
+export function shouldCreate(instance: any): void {
+  it('should create', () => expect(instance).toBeTruthy());
+}
