+++ projects/ngx-testing-extra/src/lib/core/extra-bed.ts
@@ -1,13 +1,13 @@
-import { ComponentRef, DebugElement, Type } from '@angular/core';
+import { Type } from '@angular/core';
 import { ComponentFixture, TestBed, TestModuleMetadata } from '@angular/core/testing';
+import { shouldCreate } from './expect/should-create';
+import { ExtraCb, ExtraFn, ExtraOptions } from './models';
 
 export class ExtraBed<ComponentType> {
 
   private constructor(
     private describedComponent: Type<ComponentType>,
-  ) {
-    this.import(describedComponent);
-  }
+  ) { }
 
   public static root<T>(component: Type<T>): ExtraBed<T> {
     return new ExtraBed<T>(component);
@@ -15,38 +15,32 @@ export class ExtraBed<ComponentType> {
 
   public import(imp: Type<any>): this
   public import(imps: Type<any>[]): this
-  public import(oneOrManyImport: Type<any> | Type<any>[]): this {
-    const imports: Type<any>[] = (Array.isArray(oneOrManyImport) && oneOrManyImport.length > 1)
-      ? oneOrManyImport
-      : [oneOrManyImport as Type<any>];
-    TestBed.configureTestingModule({ imports });
-    return this;
+  public import(oneOrManyImports: Type<any> | Type<any>[]): this {
+    return this.configure('imports', oneOrManyImports);
 
   }
 
   public provide(provider: any): this
   public provide(providers: any[]): this
   public provide(oneOrManyProviders: any | any[]): this {
-    const providers: any[] = (Array.isArray(oneOrManyProviders) && oneOrManyProviders.length > 1)
-      ? oneOrManyProviders
-      : [oneOrManyProviders as any];
-    TestBed.configureTestingModule({ providers });
-    return this;
+    return this.configure('providers', oneOrManyProviders);
   }
 
-  public declare(components: Type<any>[]): this {
-    TestBed.configureTestingModule({ declarations: components });
-    return this;
+  public declare(component: Type<any>): this
+  public declare(components: Type<any>[]): this
+  public declare(oneOrManyComponents: Type<any> | Type<any>[]): this {
+    return this.configure('declarations', oneOrManyComponents);
   }
 
-  public configureModule(def: TestModuleMetadata): this {
-    TestBed.configureTestingModule(def);
+  private configure(key: keyof TestModuleMetadata, itemS: any | any[]): this {
+    const defs: any[] = Array.isArray(itemS) ? itemS : [itemS];
+    TestBed.configureTestingModule({ [key]: defs });
     return this;
-  };
-
-  public inject = TestBed.inject;
+  }
 
   public async compile(): Promise<ExtraFn<ComponentType>> {
+    this.import(this.describedComponent);
+
     await TestBed.compileComponents();
 
     const fixture: ComponentFixture<ComponentType> = TestBed.createComponent(this.describedComponent);
@@ -57,7 +51,7 @@ export class ExtraBed<ComponentType> {
       componentRef: ref,
     } = fixture;
 
-    it('should create', () => expect(instance).toBeTruthy());
+    shouldCreate(instance);
 
     return (cb: ExtraCb<ComponentType>, opts?: ExtraOptions) => {
 
@@ -73,19 +67,4 @@ export class ExtraBed<ComponentType> {
       });
     };
   }
-}
-
-type MaybePromise<T> = T | Promise<T>;
-export type ExtraFn<T> = (cb: ExtraCb<T>, opts?: ExtraOptions) => jasmine.ImplementationCallback
-export type ExtraCb<T> = (tools: ExtraTools<T>) => jasmine.ImplementationCallback
-
-export interface ExtraTools<T> {
-  fixture: ComponentFixture<T>;
-  ref: ComponentRef<T>;
-  instance: T;
-  debug: DebugElement;
-}
-
-export interface ExtraOptions {
-  autoDetectChanges?: boolean;
 }
