+++ projects/ngx-testing-extra/src/lib/core/index.ts
@@ -1,2 +1,2 @@
-export { BedTools, describeComponent } from './extra-describe';
-export { ExtraFn, ExtraTools, ExtraBed } from './extra-bed';
+export * from './models';
+export { ExtraBed } from './extra-bed';
