+++ projects/ngx-testing-extra/src/lib/core/models/index.ts
@@ -1,0 +1,1 @@
+export { ExtraFn, ExtraTools, ExtraCb, ExtraOptions } from './extra.models';
